+++ src/auth/jwt_key_rotation_manager.cpp
@@ -50,7 +50,9 @@ void JWTKeyRotationManager::rotateActiveKey(const std::string &new_kid, std::opt
         std::lock_guard<std::mutex> lock(mutex_);
 
         // Enforce max_keys resource limit (new key will be added)
-        if (config_.max_keys > 0 && static_cast<int>(keys_.size()) >= config_.max_keys && keys_.find(new_kid) == keys_.end()) {
+        if (config_.max_keys > 0
+            && keys_.size() >= static_cast<std::size_t>(config_.max_keys)
+            && keys_.find(new_kid) == keys_.end()) {
             if (audit_logger_) {
                 nlohmann::json meta;
                 meta["new_kid"]   = new_kid;
